moved interfaces into types file
